requirements.txt
fastapi
fastmcp
uvicorn
screeninfo
pyautogui
pillow
pydantic
pydantic-settings
pydantic-core